requirements.txt
#该项目本地运行的环境如下：
dbutils==3.1.2
Flask==3.1.2
ftfy==6.3.1
mysql_connector_repackaged==0.3.1
numpy==2.3.5
opencv_python==4.12.0.88
packaging==25.0
Pillow==12.0.0
python-dotenv==1.2.1
regex==2025.11.3
Requests==2.32.5
torch==2.5.1
torchvision==0.20.1
